#2923: relocate ajax logic, and 'form' html wrapper
