+++ src/jsx/import/content/support-vector.jsx
@@ -20,7 +20,6 @@ import setCurrentResult from '../redux/action/current-result.jsx';
 import { setSvButton, setGotoResultsButton, setLayout } from '../redux/action/page.jsx';
 import checkValidString from '../validator/valid-string.js';
 import checkValidFloat from '../validator/valid-float.js';
-import ajaxCaller from '../general/ajax-caller.js';
 
 var SupportVector = React.createClass({
   // initial 'state properties'
@@ -32,97 +31,6 @@ var SupportVector = React.createClass({
             ajax_fail_error: null,
             ajax_fail_status: null
         };
-    },
-  // callback: get session type
-    getSessionType: function(type) {
-        return {
-            data_new: DataNewState,
-            data_append: DataAppendState,
-            model_generate: ModelGenerateState,
-            model_predict: ModelPredictState
-        }[type] || 'span';
-    },
-  // update 'state properties'
-    changeSessionType: function(event){
-      // reset value(s)
-        this.setState({
-            ajax_done_result: null,
-        });
-
-      // define sessionType
-        if (
-            event.target.value &&
-            checkValidString(event.target.value) &&
-            this.state.session_type_value != event.target.value &&
-            this.state.session_type != this.getSessionType(event.target.value)
-        ) {
-          // current component: accessed via form element update
-            const url_suffix = event.target.value.replace(/_/g, '-');
-            <Link to={'/session/' + url_suffix} replace />
-
-          // update states
-            this.setState({
-                session_type: this.getSessionType(event.target.value),
-                session_type_value: event.target.value
-            });
-        }
-
-      // update redux
-        this.props.dispatchSvButton(setSvButton({button: {submit_analysis: false}}));
-    },
-  // send form data to serverside on form submission
-    handleSubmit: function(event) {
-      // prevent page reload
-        event.preventDefault();
-
-      // local variables
-        var sessionType = this.state.session_type_value;
-        if (
-            sessionType == 'data_new' ||
-            sessionType == 'data_append' ||
-            sessionType == 'model_generate' ||
-            sessionType == 'model_predict'
-        ) {
-            const ajaxEndpoint = '/load-data';
-            var ajaxArguments = {
-                'endpoint': ajaxEndpoint,
-                'data': new FormData(this.refs.analysisForm)
-            };
-
-          // boolean to show ajax spinner
-            this.setState({display_spinner: true});
-
-          // asynchronous callback: ajax 'done' promise
-            ajaxCaller(function (asynchObject) {
-            // Append to DOM
-                if (asynchObject && asynchObject.error) {
-                    this.setState({ajax_done_error: asynchObject.error});
-                } else if (asynchObject) {
-                    this.setState({ajax_done_result: asynchObject});
-                    this.storeResults();
-                }
-                else {
-                    this.setState({ajax_done_result: null});
-                }
-            // boolean to hide ajax spinner
-                this.setState({display_spinner: false});
-            }.bind(this),
-          // asynchronous callback: ajax 'fail' promise
-            function (asynchStatus, asynchError) {
-                if (asynchStatus) {
-                    this.setState({ajax_fail_status: asynchStatus});
-                    console.log('Error Status: ' + asynchStatus);
-                }
-                if (asynchError) {
-                    this.setState({ajax_fail_error: asynchError});
-                    console.log('Error Thrown: ' + asynchError);
-                }
-            // boolean to hide ajax spinner
-                this.setState({display_spinner: false});
-            }.bind(this),
-          // pass ajax arguments
-            ajaxArguments);
-        }
     },
   // define properties before mount
     componentWillMount: function() {
@@ -240,54 +148,16 @@ var SupportVector = React.createClass({
         }
 
       // update redux store: define overall page layout
-        const action = setLayout({'layout': 'Analysis'});
+        const action = setLayout({'layout': 'analysis'});
         this.props.dispatchLayout(action);
 
-        {/* return:
-            @analysisForm, attribute is used within 'handleSubmit' callback
-            @formResult, is accessible within child component as
-                'this.props.formResult'
-        */}
         return(
-            <form onSubmit={this.handleSubmit} ref='analysisForm'>
-                <fieldset className='fieldset-session-type'>
-                    <legend>Session Type</legend>
-                    <p>Choose a session type</p>
-                    <select
-                        name='session_type'
-                        autoComplete='off'
-                        onChange={this.changeSessionType}
-                        value={this.state.session_type_value}
-                    >
-                        <option value='' defaultValue>
-                            --Select--
-                        </option>
-
-                        <option value='data_new'>
-                            New Data
-                        </option>
-
-                        <option value='data_append'>
-                            Append Data
-                        </option>
-
-
-                        <option value='model_generate'>
-                            Generate Model
-                        </option>
-
-                        <option value='model_predict'>
-                            Make Prediction
-                        </option>
-                    </select>
-                </fieldset>
-
+            <div>
                 {session}
-
                 {submitBtn}
                 {resultBtn}
                 {spinner}
-            </form>
+            </div>
         );
     },
 });
